checkpoint

+++ lib/tasks/dev.rake
@@ -3,12 +3,15 @@ namespace :dev do
   DEFAULT_PASSWORD = 123456
   DEFAULT_FILES_PATH = File.join(Rails.root, 'lib', 'tmp')
 
-  desc "Configura o ambiente..."
+  desc "Configura o ambiente de desenvolvimento"
   task setup: :environment do
+    if Rails.env.development?
       show_spinner("Apagando BD...") { %x(rails db:drop) }
       show_spinner("Criando BD...") { %x(rails db:create) }
       show_spinner("Migrando BD...") { %x(rails db:migrate) }
       show_spinner("Cadastrando o administrador padrão...") { %x(rails dev:add_default_admin) }
+    else
+      puts "Você não está em ambiente de desenvolvimento!"
     end
   end
 
@@ -19,7 +22,6 @@ namespace :dev do
       password: DEFAULT_PASSWORD,
       password_confirmation: DEFAULT_PASSWORD
     )
-    end
   end
 
   private
